@@ -16,18 +16,22 @@
   "readmeFilename": "README.md",
   "//": "Callstats.io does not work with recent versions of jsSHA (2.0.1 in particular)",
   "dependencies": {
-    "@atlassian/aui": "6.0.6",
     "@atlaskit/button": "1.0.3",
     "@atlaskit/button-group": "1.0.0",
+    "@atlaskit/dropdown-menu": "1.4.0",
+    "@atlaskit/field-text": "2.0.3",
+    "@atlaskit/icon": "6.0.0",
     "@atlaskit/modal-dialog": "1.2.4",
+    "@atlaskit/tabs": "1.2.5",
+    "@atlassian/aui": "6.0.6",
     "async": "0.9.0",
     "autosize": "1.18.13",
     "bootstrap": "3.1.1",
     "es6-iterator": "2.0.1",
     "es6-symbol": "3.1.1",
-    "i18next": "7.1.2",
+    "i18next": "8.0.0",
     "i18next-browser-languagedetector": "1.0.1",
-    "i18next-xhr-backend": "1.4.0",
+    "i18next-xhr-backend": "1.4.1",
     "jitsi-meet-logger": "jitsi/jitsi-meet-logger",
     "jquery": "2.1.4",
     "jquery-contextmenu": "2.4.3",
@@ -37,46 +41,48 @@
     "jssha": "1.5.0",
     "jws": "3.1.4",
     "lib-jitsi-meet": "jitsi/lib-jitsi-meet",
+    "lodash": "4.17.4",
     "postis": "2.2.0",
     "react": "15.4.2",
     "react-dom": "15.4.2",
-    "react-i18next": "2.2.0",
+    "react-i18next": "3.1.0",
     "react-native": "0.42.3",
     "react-native-background-timer": "1.0.0",
     "react-native-immersive": "0.0.4",
     "react-native-keep-awake": "2.0.3",
     "react-native-locale-detector": "1.0.1",
     "react-native-prompt": "1.0.0",
-    "react-native-vector-icons": "4.0.0",
+    "react-native-vector-icons": "4.0.1",
     "react-native-webrtc": "jitsi/react-native-webrtc",
-    "react-redux": "5.0.3",
+    "react-redux": "5.0.4",
     "redux": "3.6.0",
     "redux-thunk": "2.2.0",
     "retry": "0.6.1",
     "strophe": "1.2.4",
     "strophejs-plugins": "0.0.7",
+    "styled-components": "1.3.0",
     "toastr": "2.1.2",
     "url-polyfill": "github/url-polyfill",
     "xmldom": "0.1.27"
   },
   "devDependencies": {
-    "babel-core": "6.24.0",
-    "babel-eslint": "7.2.0",
-    "babel-loader": "6.4.1",
+    "babel-core": "6.24.1",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "7.0.0",
     "babel-polyfill": "6.23.0",
-    "babel-preset-es2015": "6.24.0",
-    "babel-preset-react": "6.23.0",
-    "babel-preset-stage-1": "6.22.0",
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-1": "6.24.1",
     "clean-css": "3.4.25",
-    "css-loader": "0.26.2",
-    "eslint": "3.18.0",
-    "eslint-plugin-flowtype": "2.30.3",
+    "css-loader": "0.28.0",
+    "eslint": "3.19.0",
+    "eslint-plugin-flowtype": "2.30.4",
     "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsdoc": "3.0.0",
+    "eslint-plugin-jsdoc": "3.0.2",
     "eslint-plugin-react": "6.10.3",
-    "eslint-plugin-react-native": "2.3.1",
+    "eslint-plugin-react-native": "2.3.2",
     "expose-loader": "0.7.3",
-    "file-loader": "0.10.1",
+    "file-loader": "0.11.1",
     "flow-bin": "0.38.0",
     "haste-resolver-webpack-plugin": "0.2.2",
     "imports-loader": "0.7.1",
@@ -84,8 +90,8 @@
     "json-loader": "0.5.4",
     "node-sass": "3.13.1",
     "precommit-hook": "3.0.0",
-    "string-replace-loader": "1.0.5",
-    "style-loader": "0.15.0",
+    "string-replace-loader": "1.2.0",
+    "style-loader": "0.16.1",
     "webpack": "1.14.0",
     "webpack-dev-server": "1.16.3"
   },
