@@ -56,7 +56,14 @@ export default class AbstractDialog extends Component {
         /**
          * Key to use for showing a title.
          */
-        titleKey: React.PropTypes.string
+        titleKey: React.PropTypes.string,
+
+        /**
+         * The string to use as a title instead of {@code titleKey}. If a truthy
+         * value is specified, it takes precedence over {@code titleKey} i.e.
+         * the latter is unused.
+         */
+        titleString: React.PropTypes.string
     }
 
     /**
