@@ -4,21 +4,19 @@ import { connect as reactReduxConnect } from 'react-redux';
 import { connect, disconnect } from '../../base/connection';
 import { DialogContainer } from '../../base/dialog';
 import { Container } from '../../base/react';
-import { FilmStrip } from '../../film-strip';
+import { Filmstrip } from '../../filmstrip';
 import { LargeVideo } from '../../large-video';
-import { RoomLockPrompt } from '../../room-lock';
-import { Toolbar } from '../../toolbar';
+import { setToolboxVisible, Toolbox } from '../../toolbox';
 
-import PasswordRequiredPrompt from './PasswordRequiredPrompt';
 import { styles } from './styles';
 
 /**
- * The timeout in milliseconds after which the toolbar will be hidden.
+ * The timeout in milliseconds after which the Toolbox will be hidden.
  *
  * @private
  * @type {number}
  */
-const _TOOLBAR_TIMEOUT_MS = 5000;
+const _TOOLBOX_TIMEOUT_MS = 5000;
 
 /**
  * The conference page of the mobile (i.e. React Native) application.
@@ -31,24 +29,38 @@ class Conference extends Component {
      */
     static propTypes = {
         /**
-         * The indicator which determines whether a password is required to join
-         * the conference and has not been provided yet.
+         * The handler which dispatches the (redux) action connect.
          *
          * @private
-         * @type {JitsiConference}
+         * @type {Function}
          */
-        _passwordRequired: React.PropTypes.object,
+        _onConnect: React.PropTypes.func,
 
         /**
-         * The indicator which determines whether the user has requested to lock
-         * the conference/room.
+         * The handler which dispatches the (redux) action disconnect.
          *
          * @private
-         * @type {JitsiConference}
+         * @type {Function}
          */
-        _roomLockRequested: React.PropTypes.object,
-        dispatch: React.PropTypes.func
-    }
+        _onDisconnect: React.PropTypes.func,
+
+        /**
+         * The handler which dispatches the (redux) action setToolboxVisible to
+         * show/hide the Toolbox.
+         *
+         * @private
+         * @type {boolean}
+         */
+        _setToolboxVisible: React.PropTypes.func,
+
+        /**
+         * The indicator which determines whether the Toolbox is visible.
+         *
+         * @private
+         * @type {boolean}
+         */
+        _toolboxVisible: React.PropTypes.bool
+    };
 
     /**
      * Initializes a new Conference instance.
@@ -59,29 +71,27 @@ class Conference extends Component {
     constructor(props) {
         super(props);
 
-        this.state = { toolbarVisible: true };
-
         /**
          * The numerical ID of the timeout in milliseconds after which the
-         * toolbar will be hidden. To be used with
+         * Toolbox will be hidden. To be used with
          * {@link WindowTimers#clearTimeout()}.
          *
          * @private
          */
-        this._toolbarTimeout = undefined;
+        this._toolboxTimeout = undefined;
 
         // Bind event handlers so they are only bound once for every instance.
         this._onClick = this._onClick.bind(this);
     }
 
     /**
-     * Inits the toolbar timeout after the component is initially rendered.
+     * Inits the Toolbox timeout after the component is initially rendered.
      *
      * @inheritdoc
      * returns {void}
      */
     componentDidMount() {
-        this._setToolbarTimeout(this.state.toolbarVisible);
+        this._setToolboxTimeout(this.props._toolboxVisible);
     }
 
     /**
@@ -91,20 +101,20 @@ class Conference extends Component {
      * @returns {void}
      */
     componentWillMount() {
-        this.props.dispatch(connect());
+        this.props._onConnect();
     }
 
     /**
      * Destroys connection, conference and local tracks when conference screen
-     * is left. Clears {@link #_toolbarTimeout} before the component unmounts.
+     * is left. Clears {@link #_toolboxTimeout} before the component unmounts.
      *
      * @inheritdoc
      * @returns {void}
      */
     componentWillUnmount() {
-        this._clearToolbarTimeout();
+        this._clearToolboxTimeout();
 
-        this.props.dispatch(disconnect());
+        this.props._onDisconnect();
     }
 
     /**
@@ -114,8 +124,6 @@ class Conference extends Component {
      * @returns {ReactElement}
      */
     render() {
-        const toolbarVisible = this.state.toolbarVisible;
-
         return (
             <Container
                 onClick = { this._onClick }
@@ -123,112 +131,104 @@ class Conference extends Component {
                 touchFeedback = { false }>
 
                 <LargeVideo />
-                <Toolbar visible = { toolbarVisible } />
-                <FilmStrip visible = { !toolbarVisible } />
+
+                <Toolbox />
+                <Filmstrip />
 
                 <DialogContainer />
-
-                {
-                    this._renderPrompt()
-                }
             </Container>
         );
     }
 
     /**
-     * Clears {@link #_toolbarTimeout} if any.
+     * Clears {@link #_toolboxTimeout} if any.
      *
      * @private
      * @returns {void}
      */
-    _clearToolbarTimeout() {
-        if (this._toolbarTimeout) {
-            clearTimeout(this._toolbarTimeout);
-            this._toolbarTimeout = undefined;
+    _clearToolboxTimeout() {
+        if (this._toolboxTimeout) {
+            clearTimeout(this._toolboxTimeout);
+            this._toolboxTimeout = undefined;
         }
     }
 
     /**
-     * Changes the value of the toolbarVisible state, thus allowing us to
-     * 'switch' between toolbar and filmstrip views and change the visibility of
-     * the above.
+     * Changes the value of the toolboxVisible state, thus allowing us to switch
+     * between Toolbox and Filmstrip and change their visibility.
      *
      * @private
      * @returns {void}
      */
     _onClick() {
-        const toolbarVisible = !this.state.toolbarVisible;
+        const toolboxVisible = !this.props._toolboxVisible;
 
-        this.setState({ toolbarVisible });
-
-        this._setToolbarTimeout(toolbarVisible);
+        this.props._setToolboxVisible(toolboxVisible);
+        this._setToolboxTimeout(toolboxVisible);
     }
 
     /**
-     * Renders a prompt if a password is required to join the conference.
+     * Triggers the default Toolbox timeout.
      *
-     * @private
-     * @returns {ReactElement}
-     */
-    _renderPasswordRequiredPrompt() {
-        const required = this.props._passwordRequired;
-
-        if (required) {
-            return (
-                <PasswordRequiredPrompt conference = { required } />
-            );
-        }
-
-        return null;
-    }
-
-    /**
-     * Renders a prompt if necessary such as when a password is required to join
-     * the conference or the user has requested to lock the conference/room.
-     *
-     * @private
-     * @returns {ReactElement}
-     */
-    _renderPrompt() {
-        return (
-            this._renderPasswordRequiredPrompt()
-                || this._renderRoomLockPrompt()
-        );
-    }
-
-    /**
-     * Renders a prompt if the user has requested to lock the conference/room.
-     *
-     * @private
-     * @returns {ReactElement}
-     */
-    _renderRoomLockPrompt() {
-        const requested = this.props._roomLockRequested;
-
-        if (requested) {
-            return (
-                <RoomLockPrompt conference = { requested } />
-            );
-        }
-
-        return null;
-    }
-
-    /**
-     * Triggers the default toolbar timeout.
-     *
-     * @param {boolean} toolbarVisible - Indicates if the toolbar is currently
-     * visible.
+     * @param {boolean} toolboxVisible - Indicates whether the Toolbox is
+     * currently visible.
      * @private
      * @returns {void}
      */
-    _setToolbarTimeout(toolbarVisible) {
-        this._clearToolbarTimeout();
-        if (toolbarVisible) {
-            this._toolbarTimeout
-                = setTimeout(this._onClick, _TOOLBAR_TIMEOUT_MS);
+    _setToolboxTimeout(toolboxVisible) {
+        this._clearToolboxTimeout();
+        if (toolboxVisible) {
+            this._toolboxTimeout
+                = setTimeout(this._onClick, _TOOLBOX_TIMEOUT_MS);
         }
     }
+}
+
+/**
+ * Maps dispatching of some action to React component props.
+ *
+ * @param {Function} dispatch - Redux action dispatcher.
+ * @private
+ * @returns {{
+ *     _onConnect: Function,
+ *     _onDisconnect: Function,
+ *     _setToolboxVisible: Function
+ * }}
+ */
+function _mapDispatchToProps(dispatch) {
+    return {
+        /**
+         * Dispatched an action connecting to the conference.
+         *
+         * @returns {Object} Dispatched action.
+         * @private
+         */
+        _onConnect() {
+            return dispatch(connect());
+        },
+
+        /**
+         * Dispatches an action disconnecting from the conference.
+         *
+         * @returns {Object} Dispatched action.
+         * @private
+         */
+        _onDisconnect() {
+            return dispatch(disconnect());
+        },
+
+        /**
+         * Dispatches an action changing the visiblity of the Toolbox.
+         *
+         * @param {boolean} visible - True to show the Toolbox or false to hide
+         * it.
+         * @returns {Object} Dispatched action.
+         * @private
+         */
+        _setToolboxVisible(visible: boolean) {
+            return dispatch(setToolboxVisible(visible));
+        }
+    };
 }
 
 /**
@@ -237,29 +237,20 @@ class Conference extends Component {
  * @param {Object} state - The Redux state.
  * @private
  * @returns {{
- *     _passwordRequired: boolean
+ *     _toolboxVisible: boolean
  * }}
  */
 function _mapStateToProps(state) {
     return {
         /**
-         * The indicator which determines whether a password is required to join
-         * the conference and has not been provided yet.
+         * The indicator which determines whether the Toolbox is visible.
          *
          * @private
-         * @type {JitsiConference}
+         * @type {boolean}
          */
-        _passwordRequired: state['features/base/conference'].passwordRequired,
-
-        /**
-         * The indicator which determines whether the user has requested to lock
-         * the conference/room.
-         *
-         * @private
-         * @type {JitsiConference}
-         */
-        _roomLockRequested: state['features/room-lock'].requested
+        _toolboxVisible: state['features/toolbox'].visible
     };
 }
 
-export default reactReduxConnect(_mapStateToProps)(Conference);
+export default reactReduxConnect(_mapStateToProps, _mapDispatchToProps)(
+        Conference);
