@@ -2,7 +2,6 @@
 const logger = require("jitsi-meet-logger").getLogger(__filename);
 
 import {openConnection} from './connection';
-import Invite from './modules/UI/invite/Invite';
 import ContactList from './modules/UI/side_pannels/contactlist/ContactList';
 
 import AuthHandler from './modules/UI/authentication/AuthHandler';
@@ -20,25 +19,33 @@ import analytics from './modules/analytics/analytics';
 
 import EventEmitter from "events";
 
+import { showDesktopSharingButton } from './react/features/toolbox';
+
 import {
     AVATAR_ID_COMMAND,
     AVATAR_URL_COMMAND,
     conferenceFailed,
     conferenceJoined,
     conferenceLeft,
-    EMAIL_COMMAND
+    EMAIL_COMMAND,
+    lockStateChanged
 } from './react/features/base/conference';
+import {
+    updateDeviceList
+} from './react/features/base/devices';
 import {
     isFatalJitsiConnectionError
 } from './react/features/base/lib-jitsi-meet';
 import {
-    changeParticipantAvatarID,
-    changeParticipantAvatarURL,
-    changeParticipantEmail,
+    localParticipantRoleChanged,
     participantJoined,
     participantLeft,
-    participantRoleChanged
+    participantRoleChanged,
+    participantUpdated
 } from './react/features/base/participants';
+import {
+    showDesktopPicker
+} from  './react/features/desktop-picker';
 import {
     mediaPermissionPromptVisibilityChanged,
     suspendDetected
@@ -57,7 +64,10 @@ const ConnectionQualityEvents = JitsiMeetJS.events.connectionQuality;
 
 const eventEmitter = new EventEmitter();
 
-let room, connection, localAudio, localVideo;
+let room;
+let connection;
+let localAudio, localVideo;
+let initialAudioMutedState = false, initialVideoMutedState = false;
 
 /**
  * Indicates whether extension external installation is in progress or not.
@@ -65,6 +75,16 @@ let room, connection, localAudio, localVideo;
 let DSExternalInstallationInProgress = false;
 
 import {VIDEO_CONTAINER_TYPE} from "./modules/UI/videolayout/VideoContainer";
+
+/*
+ * Logic to open a desktop picker put on the window global for
+ * lib-jitsi-meet to detect and invoke
+ */
+window.JitsiMeetScreenObtainer = {
+    openDesktopPicker(onSourceChoose) {
+        APP.store.dispatch(showDesktopPicker(onSourceChoose));
+    }
+};
 
 /**
  * Known custom conference commands.
@@ -160,7 +180,11 @@ function createInitialLocalTracksAndConnect(roomName) {
  * @param command the command
  * @param {string} value new value
  */
-function sendData (command, value) {
+function sendData(command, value) {
+    if(!room) {
+        return;
+    }
+
     room.removeCommand(command);
     room.sendCommand(command, {value: value});
 }
@@ -193,7 +217,7 @@ function _setupLocalParticipantProperties() {
  * @param {string} id user id
  * @returns {string?} user nickname or undefined if user is unknown.
  */
-function getDisplayName (id) {
+function getDisplayName(id) {
     if (APP.conference.isLocalId(id)) {
         return APP.settings.getDisplayName();
     }
@@ -210,7 +234,7 @@ function getDisplayName (id) {
  * @param {boolean} userInteraction - indicates if this local audio mute was a
  * result of user interaction
  */
-function muteLocalAudio (muted) {
+function muteLocalAudio(muted) {
     muteLocalMedia(localAudio, muted, 'Audio');
 }
 
@@ -230,7 +254,7 @@ function muteLocalMedia(localMedia, muted, localMediaTypeString) {
  * Mute or unmute local video stream if it exists.
  * @param {boolean} muted if video stream should be muted or unmuted.
  */
-function muteLocalVideo (muted) {
+function muteLocalVideo(muted) {
     muteLocalMedia(localVideo, muted, 'Video');
 }
 
@@ -315,7 +339,7 @@ function assignWindowLocationPathname(pathname) {
  *      for gUM permission prompt
  * @returns {Promise<JitsiLocalTrack[]>}
  */
-function createLocalTracks (options, checkForPermissionPrompt) {
+function createLocalTracks(options, checkForPermissionPrompt) {
     options || (options = {});
 
     return JitsiMeetJS
@@ -348,28 +372,10 @@ function createLocalTracks (options, checkForPermissionPrompt) {
         });
 }
 
-/**
- * Changes the display name for the local user
- * @param nickname {string} the new display name
- */
-function changeLocalDisplayName(nickname = '') {
-    const formattedNickname
-        = nickname.trim().substr(0, MAX_DISPLAY_NAME_LENGTH);
-
-    if (formattedNickname === APP.settings.getDisplayName()) {
-        return;
-    }
-
-    APP.settings.setDisplayName(formattedNickname);
-    room.setDisplayName(formattedNickname);
-    APP.UI.changeDisplayName(APP.conference.getMyUserId(), formattedNickname);
-}
-
 class ConferenceConnector {
-    constructor(resolve, reject, invite) {
+    constructor(resolve, reject) {
         this._resolve = resolve;
         this._reject = reject;
-        this._invite = invite;
         this.reconnectTimeout = null;
         room.on(ConferenceEvents.CONFERENCE_JOINED,
             this._handleConferenceJoined.bind(this));
@@ -387,10 +393,6 @@ class ConferenceConnector {
         logger.error('CONFERENCE FAILED:', err, ...params);
         APP.UI.hideRingOverLay();
         switch (err) {
-            // room is locked by the password
-        case ConferenceErrors.PASSWORD_REQUIRED:
-            APP.UI.emitEvent(UIEvents.PASSWORD_REQUIRED);
-            break;
 
         case ConferenceErrors.CONNECTION_ERROR:
             {
@@ -407,15 +409,15 @@ class ConferenceConnector {
             break;
 
             // not enough rights to create conference
-        case ConferenceErrors.AUTHENTICATION_REQUIRED:
-            // schedule reconnect to check if someone else created the room
-            this.reconnectTimeout = setTimeout(function () {
-                room.join();
-            }, 5000);
+        case ConferenceErrors.AUTHENTICATION_REQUIRED: {
+                // Schedule reconnect to check if someone else created the room.
+                this.reconnectTimeout = setTimeout(() => room.join(), 5000);
 
-            // notify user that auth is required
-            AuthHandler.requireAuth(
-                room, this._invite.getRoomLocker().password);
+                const { password }
+                    = APP.store.getState()['features/base/conference'];
+
+                AuthHandler.requireAuth(room, password);
+            }
             break;
 
         case ConferenceErrors.RESERVATION_ERROR:
@@ -582,6 +584,12 @@ export default {
                 analytics.init();
                 return createInitialLocalTracksAndConnect(options.roomName);
             }).then(([tracks, con]) => {
+                tracks.forEach(track => {
+                    if((track.isAudioTrack() && initialAudioMutedState)
+                        || (track.isVideoTrack() && initialVideoMutedState)) {
+                        track.mute();
+                    }
+                });
                 logger.log('initialized with %s local tracks', tracks.length);
                 con.addEventListener(
                     ConnectionEvents.CONNECTION_FAILED,
@@ -589,6 +597,12 @@ export default {
                 APP.connection = connection = con;
                 this.isDesktopSharingEnabled =
                     JitsiMeetJS.isDesktopSharingEnabled();
+                eventEmitter.emit(
+                    JitsiMeetConferenceEvents.DESKTOP_SHARING_ENABLED_CHANGED,
+                    this.isDesktopSharingEnabled);
+
+                APP.store.dispatch(showDesktopSharingButton());
+
                 APP.remoteControl.init();
                 this._createRoom(tracks);
 
@@ -615,8 +629,7 @@ export default {
                 // XXX The API will take care of disconnecting from the XMPP
                 // server (and, thus, leaving the room) on unload.
                 return new Promise((resolve, reject) => {
-                    (new ConferenceConnector(
-                        resolve, reject, this.invite)).connect();
+                    (new ConferenceConnector(resolve, reject)).connect();
                 });
         });
     },
@@ -625,14 +638,14 @@ export default {
      * @param {string} id id to check
      * @returns {boolean}
      */
-    isLocalId (id) {
+    isLocalId(id) {
         return this.getMyUserId() === id;
     },
     /**
      * Simulates toolbar button click for audio mute. Used by shortcuts and API.
      * @param mute true for mute and false for unmute.
      */
-    muteAudio (mute) {
+    muteAudio(mute) {
         muteLocalAudio(mute);
     },
     /**
@@ -643,36 +656,51 @@ export default {
         return this.audioMuted;
     },
     /**
-     * Simulates toolbar button click for audio mute. Used by shortcuts and API.
+     * Simulates toolbar button click for audio mute. Used by shortcuts
+     * and API.
+     * @param {boolean} force - If the track is not created, the operation
+     * will be executed after the track is created. Otherwise the operation
+     * will be ignored.
      */
-    toggleAudioMuted () {
+    toggleAudioMuted(force = false) {
+        if(!localAudio && force) {
+            initialAudioMutedState = !initialAudioMutedState;
+            return;
+        }
         this.muteAudio(!this.audioMuted);
     },
     /**
      * Simulates toolbar button click for video mute. Used by shortcuts and API.
      * @param mute true for mute and false for unmute.
      */
-    muteVideo (mute) {
+    muteVideo(mute) {
         muteLocalVideo(mute);
     },
     /**
      * Simulates toolbar button click for video mute. Used by shortcuts and API.
+     * @param {boolean} force - If the track is not created, the operation
+     * will be executed after the track is created. Otherwise the operation
+     * will be ignored.
      */
-    toggleVideoMuted () {
+    toggleVideoMuted(force = false) {
+        if(!localVideo && force) {
+            initialVideoMutedState = !initialVideoMutedState;
+            return;
+        }
         this.muteVideo(!this.videoMuted);
     },
     /**
      * Retrieve list of conference participants (without local user).
      * @returns {JitsiParticipant[]}
      */
-    listMembers () {
+    listMembers() {
         return room.getParticipants();
     },
     /**
      * Retrieve list of ids of conference participants (without local user).
      * @returns {string[]}
      */
-    listMembersIds () {
+    listMembersIds() {
         return room.getParticipants().map(p => p.getId());
     },
     /**
@@ -680,7 +708,7 @@ export default {
      * @id id to search for participant
      * @return {boolean} whether the participant is moderator
      */
-    isParticipantModerator (id) {
+    isParticipantModerator(id) {
         let user = room.getParticipantById(id);
         return user && user.isModerator();
     },
@@ -688,10 +716,10 @@ export default {
      * Check if SIP is supported.
      * @returns {boolean}
      */
-    sipGatewayEnabled () {
+    sipGatewayEnabled() {
         return room.isSIPCallingSupported();
     },
-    get membersCount () {
+    get membersCount() {
         return room.getParticipants().length + 1;
     },
     /**
@@ -701,7 +729,7 @@ export default {
      * @returns true if the callstats integration is enabled, otherwise returns
      * false.
      */
-    isCallstatsEnabled () {
+    isCallstatsEnabled() {
         return room && room.isCallstatsEnabled();
     },
     /**
@@ -711,7 +739,7 @@ export default {
      * user feedback
      * @param detailedFeedback detailed feedback from the user. Not yet used
      */
-    sendFeedback (overallFeedback, detailedFeedback) {
+    sendFeedback(overallFeedback, detailedFeedback) {
         return room.sendFeedback (overallFeedback, detailedFeedback);
     },
 
@@ -729,15 +757,15 @@ export default {
     /**
      * Returns the connection times stored in the library.
      */
-    getConnectionTimes () {
+    getConnectionTimes() {
         return this._room.getConnectionTimes();
     },
     // used by torture currently
-    isJoined () {
+    isJoined() {
         return this._room
             && this._room.isJoined();
     },
-    getConnectionState () {
+    getConnectionState() {
         return this._room
             && this._room.getConnectionState();
     },
@@ -746,7 +774,7 @@ export default {
      * @return {string|null} ICE connection state or <tt>null</tt> if there's no
      * P2P connection
      */
-    getP2PConnectionState () {
+    getP2PConnectionState() {
         return this._room
             && this._room.getP2PConnectionState();
     },
@@ -754,7 +782,7 @@ export default {
      * Starts P2P (for tests only)
      * @private
      */
-    _startP2P () {
+    _startP2P() {
         try {
             this._room && this._room.startP2PSession();
         } catch (error) {
@@ -766,7 +794,7 @@ export default {
      * Stops P2P (for tests only)
      * @private
      */
-    _stopP2P () {
+    _stopP2P() {
         try {
             this._room && this._room.stopP2PSession();
         } catch (error) {
@@ -781,7 +809,7 @@ export default {
      * @returns {boolean} true if the connection is in interrupted state or
      * false otherwise.
      */
-    isConnectionInterrupted () {
+    isConnectionInterrupted() {
         return this._room.isConnectionInterrupted();
     },
     /**
@@ -792,20 +820,20 @@ export default {
      * @returns {JitsiParticipant|null} participant instance for given id or
      * null if not found.
      */
-    getParticipantById (id) {
+    getParticipantById(id) {
         return room ? room.getParticipantById(id) : null;
     },
     /**
-     * Checks whether the user identified by given id is currently connected.
+     * Get participant connection status for the participant.
      *
      * @param {string} id participant's identifier(MUC nickname)
      *
-     * @returns {boolean|null} true if participant's connection is ok or false
-     * if the user is having connectivity issues.
+     * @returns {ParticipantConnectionStatus|null} the status of the participant
+     * or null if no such participant is found or participant is the local user.
      */
-    isParticipantConnectionActive (id) {
+    getParticipantConnectionStatus(id) {
         let participant = this.getParticipantById(id);
-        return participant ? participant.isConnectionActive() : null;
+        return participant ? participant.getConnectionStatus() : null;
     },
     /**
      * Gets the display name foe the <tt>JitsiParticipant</tt> identified by
@@ -816,7 +844,7 @@ export default {
      * @return {string} the participant's display name or the default string if
      * absent.
      */
-    getParticipantDisplayName (id) {
+    getParticipantDisplayName(id) {
         let displayName = getDisplayName(id);
         if (displayName) {
             return displayName;
@@ -829,7 +857,7 @@ export default {
             }
         }
     },
-    getMyUserId () {
+    getMyUserId() {
         return this._room
             && this._room.myUserId();
     },
@@ -855,7 +883,7 @@ export default {
      * @param id the id for the user audio level to return (the id value is
      *          returned for the participant using getMyUserId() method)
      */
-    getPeerSSRCAudioLevel (id) {
+    getPeerSSRCAudioLevel(id) {
         return this.audioLevelsMap[id];
     },
     /**
@@ -875,7 +903,7 @@ export default {
     },
     // end used by torture
 
-    getLogs () {
+    getLogs() {
         return room.getLogs();
     },
 
@@ -884,7 +912,7 @@ export default {
      * debugging.
      * @param filename (optional) specify target filename
      */
-    saveLogs (filename = 'meetlog.json') {
+    saveLogs(filename = 'meetlog.json') {
         // this can be called from console and will not have reference to this
         // that's why we reference the global var
         let logs = APP.conference.getLogs();
@@ -955,7 +983,6 @@ export default {
         room = connection.initJitsiConference(APP.conference.roomName,
             this._getConferenceOptions());
         this._setLocalAudioVideoStreams(localTracks);
-        this.invite = new Invite(room);
         this._room = room; // FIXME do not use this
 
         _setupLocalParticipantProperties();
@@ -994,13 +1021,13 @@ export default {
     },
 
 
-  /**
-   * Start using provided video stream.
-   * Stops previous video stream.
-   * @param {JitsiLocalTrack} [stream] new stream to use or null
-   * @returns {Promise}
-   */
-    useVideoStream (newStream) {
+    /**
+     * Start using provided video stream.
+     * Stops previous video stream.
+     * @param {JitsiLocalTrack} [stream] new stream to use or null
+     * @returns {Promise}
+     */
+    useVideoStream(newStream) {
         return room.replaceTrack(localVideo, newStream)
             .then(() => {
                 // We call dispose after doing the replace because
@@ -1035,7 +1062,7 @@ export default {
      * @param {JitsiLocalTrack} [stream] new stream to use or null
      * @returns {Promise}
      */
-    useAudioStream (newStream) {
+    useAudioStream(newStream) {
         return room.replaceTrack(localAudio, newStream)
             .then(() => {
                 // We call dispose after doing the replace because
@@ -1058,8 +1085,9 @@ export default {
             });
     },
 
+
     videoSwitchInProgress: false,
-    toggleScreenSharing (shareScreen = !this.isSharingScreen) {
+    toggleScreenSharing(shareScreen = !this.isSharingScreen) {
         if (this.videoSwitchInProgress) {
             logger.warn("Switch in progress.");
             return;
@@ -1182,7 +1210,7 @@ export default {
     /**
      * Setup interaction between conference and UI.
      */
-    _setupListeners () {
+    _setupListeners() {
         // add local streams when joined to the conference
         room.on(ConferenceEvents.CONFERENCE_JOINED, () => {
             APP.store.dispatch(conferenceJoined(room));
@@ -1232,14 +1260,18 @@ export default {
 
 
         room.on(ConferenceEvents.USER_ROLE_CHANGED, (id, role) => {
-            APP.store.dispatch(participantRoleChanged(id, role));
             if (this.isLocalId(id)) {
                 logger.info(`My role changed, new role: ${role}`);
+
+                APP.store.dispatch(localParticipantRoleChanged(role));
+
                 if (this.isModerator !== room.isModerator()) {
                     this.isModerator = room.isModerator();
                     APP.UI.updateLocalRole(room.isModerator());
                 }
             } else {
+                APP.store.dispatch(participantRoleChanged(id, role));
+
                 let user = room.getParticipantById(id);
                 if (user) {
                     APP.UI.updateUserRole(user);
@@ -1312,8 +1344,8 @@ export default {
 
         room.on(
             ConferenceEvents.PARTICIPANT_CONN_STATUS_CHANGED,
-            (id, isActive) => {
-                APP.UI.participantConnectionStatusChanged(id, isActive);
+            id => {
+                APP.UI.participantConnectionStatusChanged(id);
         });
         room.on(ConferenceEvents.DOMINANT_SPEAKER_CHANGED, (id) => {
             if (this.isLocalId(id)) {
@@ -1336,10 +1368,15 @@ export default {
             room.on(ConferenceEvents.CONNECTION_RESTORED, () => {
                 APP.UI.markVideoInterrupted(false);
             });
-            room.on(ConferenceEvents.MESSAGE_RECEIVED, (id, text, ts) => {
+            room.on(ConferenceEvents.MESSAGE_RECEIVED, (id, body, ts) => {
                 let nick = getDisplayName(id);
-                APP.API.notifyReceivedChatMessage(id, nick, text, ts);
-                APP.UI.addMessage(id, nick, text, ts);
+                APP.API.notifyReceivedChatMessage({
+                    id,
+                    nick,
+                    body,
+                    ts
+                });
+                APP.UI.addMessage(id, nick, body, ts);
             });
             APP.UI.addListener(UIEvents.MESSAGE_CREATED, (message) => {
                 APP.API.notifySendingChatMessage(message);
@@ -1408,6 +1445,10 @@ export default {
             APP.API.notifyDisplayNameChanged(id, formattedDisplayName);
             APP.UI.changeDisplayName(id, formattedDisplayName);
         });
+
+        room.on(
+            ConferenceEvents.LOCK_STATE_CHANGED,
+            (...args) => APP.store.dispatch(lockStateChanged(room, ...args)));
 
         room.on(ConferenceEvents.PARTICIPANT_PROPERTY_CHANGED,
                 (participant, name, oldValue, newValue) => {
@@ -1486,7 +1527,10 @@ export default {
 
         APP.UI.addListener(UIEvents.EMAIL_CHANGED, this.changeLocalEmail);
         room.addCommandListener(this.commands.defaults.EMAIL, (data, from) => {
-            APP.store.dispatch(changeParticipantEmail(from, data.value));
+            APP.store.dispatch(participantUpdated({
+                id: from,
+                email: data.value
+            }));
             APP.UI.setUserEmail(from, data.value);
         });
 
@@ -1494,18 +1538,25 @@ export default {
             this.commands.defaults.AVATAR_URL,
             (data, from) => {
                 APP.store.dispatch(
-                    changeParticipantAvatarURL(from, data.value));
+                    participantUpdated({
+                        id: from,
+                        avatarURL: data.value
+                    }));
                 APP.UI.setUserAvatarUrl(from, data.value);
         });
 
         room.addCommandListener(this.commands.defaults.AVATAR_ID,
             (data, from) => {
                 APP.store.dispatch(
-                    changeParticipantAvatarID(from, data.value));
+                    participantUpdated({
+                        id: from,
+                        avatarID: data.value
+                    }));
                 APP.UI.setUserAvatarID(from, data.value);
             });
 
-        APP.UI.addListener(UIEvents.NICKNAME_CHANGED, changeLocalDisplayName);
+        APP.UI.addListener(UIEvents.NICKNAME_CHANGED,
+            this.changeLocalDisplayName.bind(this));
 
         APP.UI.addListener(UIEvents.START_MUTED_CHANGED,
             (startAudioMuted, startVideoMuted) => {
@@ -1612,7 +1663,6 @@ export default {
                 })
                 .catch((err) => {
                     APP.UI.showDeviceErrorDialog(null, err);
-                    APP.UI.setSelectedCameraFromSettings();
                 });
             }
         );
@@ -1634,7 +1684,6 @@ export default {
                 })
                 .catch((err) => {
                     APP.UI.showDeviceErrorDialog(err, null);
-                    APP.UI.setSelectedMicFromSettings();
                 });
             }
         );
@@ -1650,7 +1699,6 @@ export default {
                         logger.warn('Failed to change audio output device. ' +
                             'Default or previously set audio output device ' +
                             'will be used instead.', err);
-                        APP.UI.setSelectedAudioOutputFromSettings();
                     });
             }
         );
@@ -1746,8 +1794,8 @@ export default {
                 }
 
                 mediaDeviceHelper.setCurrentMediaDevices(devices);
-
                 APP.UI.onAvailableDevicesChanged(devices);
+                APP.store.dispatch(updateDeviceList(devices));
             });
 
             this.deviceChangeListener = (devices) =>
@@ -1880,7 +1928,7 @@ export default {
      * @param {boolean} [requestFeedback=false] if user feedback should be
      * requested
      */
-    hangup (requestFeedback = false) {
+    hangup(requestFeedback = false) {
         eventEmitter.emit(JitsiMeetConferenceEvents.BEFORE_HANGUP);
         APP.UI.hideRingOverLay();
         let requestFeedbackPromise = requestFeedback
@@ -1912,10 +1960,17 @@ export default {
         if (email === APP.settings.getEmail()) {
             return;
         }
-        APP.store.dispatch(changeParticipantEmail(room.myUserId(), email));
+
+        const localId = room ? room.myUserId() : undefined;
+
+        APP.store.dispatch(participantUpdated({
+            id: localId,
+            local: true,
+            email
+        }));
 
         APP.settings.setEmail(email);
-        APP.UI.setUserEmail(room.myUserId(), email);
+        APP.UI.setUserEmail(localId, email);
         sendData(commands.EMAIL, email);
     },
 
@@ -1929,10 +1984,17 @@ export default {
         if (url === APP.settings.getAvatarUrl()) {
             return;
         }
-        APP.store.dispatch(changeParticipantAvatarURL(room.myUserId(), url));
+
+        const localId = room ? room.myUserId() : undefined;
+
+        APP.store.dispatch(participantUpdated({
+            id: localId,
+            local: true,
+            avatarURL: url
+        }));
 
         APP.settings.setAvatarUrl(url);
-        APP.UI.setUserAvatarUrl(room.myUserId(), url);
+        APP.UI.setUserAvatarUrl(localId, url);
         sendData(commands.AVATAR_URL, url);
     },
 
@@ -1944,7 +2006,7 @@ export default {
      * @throws NetworkError or InvalidStateError or Error if the operation
      * fails.
      */
-    sendEndpointMessage (to, payload) {
+    sendEndpointMessage(to, payload) {
         room.sendEndpointMessage(to, payload);
     },
 
@@ -1953,7 +2015,7 @@ export default {
      * @param {String} eventName the name of the event
      * @param {Function} listener the listener.
      */
-    addListener (eventName, listener) {
+    addListener(eventName, listener) {
         eventEmitter.addListener(eventName, listener);
     },
 
@@ -1963,7 +2025,7 @@ export default {
      * listener
      * @param {Function} listener the listener.
      */
-    removeListener (eventName, listener) {
+    removeListener(eventName, listener) {
         eventEmitter.removeListener(eventName, listener);
     },
 
@@ -1976,7 +2038,25 @@ export default {
      * is currently in the last N set or if there's no last N set at this point
      * and {false} otherwise
      */
-    isInLastN (participantId) {
+    isInLastN(participantId) {
         return room.isInLastN(participantId);
+    },
+    /**
+     * Changes the display name for the local user
+     * @param nickname {string} the new display name
+     */
+    changeLocalDisplayName(nickname = '') {
+        const formattedNickname
+            = nickname.trim().substr(0, MAX_DISPLAY_NAME_LENGTH);
+
+        if (formattedNickname === APP.settings.getDisplayName()) {
+            return;
+        }
+
+        APP.settings.setDisplayName(formattedNickname);
+        if (room) {
+            room.setDisplayName(formattedNickname);
+            APP.UI.changeDisplayName(this.getMyUserId(), formattedNickname);
+        }
     }
 };
