@@ -1,12 +1,14 @@
-/* global $, APP, AJS, interfaceConfig, JitsiMeetJS */
-
+/* global $, APP, AJS, interfaceConfig */
 import { LANGUAGES } from "../../../../react/features/base/i18n";
+import { openDeviceSelectionDialog }
+    from '../../../../react/features/device-selection';
 
 import UIUtil from "../../util/UIUtil";
 import UIEvents from "../../../../service/UI/UIEvents";
-import Settings from '../../../settings/Settings';
 
 const sidePanelsContainerId = 'sideToolbarContainer';
+const deviceSelectionButtonClasses
+    = 'button-control button-control_primary button-control_full-width';
 const htmlStr = `
     <div id="settings_container" class="sideToolbarContainer__inner">
         <div class="title" data-i18n="settings.title"></div>
@@ -19,17 +21,11 @@ const htmlStr = `
                 <div id="deviceOptionsTitle" class="subTitle hide" 
                     data-i18n="settings.audioVideo"></div>
                 <div class="sideToolbarBlock first">
-                    <label class="first" data-i18n="settings.selectCamera">
-                    </label>
-                    <select id="selectCamera"></select>
-                </div>
-                <div class="sideToolbarBlock">
-                    <label data-i18n="settings.selectMic"></label>
-                    <select id="selectMic"></select>
-                </div>
-                <div class="sideToolbarBlock">
-                    <label data-i18n="settings.selectAudioOutput"></label>
-                    <select id="selectAudioOutput"></select>
+                    <button
+                        class="${deviceSelectionButtonClasses}"
+                        data-i18n="deviceSelection.deviceSettings"
+                        id="deviceSelection"
+                        type="button"></button>
                 </div>
             </div>
             <div id="moderatorOptionsWrapper" class="hide">
@@ -90,40 +86,6 @@ function generateLanguagesOptions(items, currentLang) {
 }
 
 /**
- * Generate html select options for available physical devices.
- *
- * @param {{ deviceId, label }[]} items available devices
- * @param {string} [selectedId] id of selected device
- * @param {boolean} permissionGranted if permission to use selected device type
- *      is granted
- * @returns {string}
- */
-function generateDevicesOptions(items, selectedId, permissionGranted) {
-    if (!permissionGranted && items.length) {
-        return '<option data-i18n="settings.noPermission"></option>';
-    }
-
-    var options = items.map(function (item) {
-        let attrs = {
-            value: item.deviceId
-        };
-
-        if (item.deviceId === selectedId) {
-            attrs.selected = 'selected';
-        }
-
-        let attrsStr = UIUtil.attrsToString(attrs);
-        return `<option ${attrsStr}>${item.label}</option>`;
-    });
-
-    if (!items.length) {
-        options.unshift('<option data-i18n="settings.noDevice"></option>');
-    }
-
-    return options.join('');
-}
-
-/**
  * Replace html select element to select2 custom dropdown
  *
  * @param {jQueryElement} $el native select element
@@ -179,13 +141,9 @@ export default {
         if (UIUtil.isSettingEnabled('devices')) {
             const wrapperId = 'deviceOptionsWrapper';
 
-            JitsiMeetJS.mediaDevices.isDeviceListAvailable()
-                .then((isDeviceListAvailable) => {
-                    if (isDeviceListAvailable &&
-                        JitsiMeetJS.mediaDevices.isDeviceChangeAvailable()) {
-                        this._initializeDeviceSelectionSettings(emitter);
-                    }
-                });
+            $('#deviceSelection').on('click', () =>
+                APP.store.dispatch(openDeviceSelectionDialog()));
+
             // Only show the subtitle if this isn't the only setting section.
             if (interfaceConfig.SETTINGS_SECTIONS.length > 1)
                 UIUtil.setVisible("deviceOptionsTitle", true);
@@ -217,30 +175,6 @@ export default {
 
             UIUtil.setVisible(wrapperId, true);
         }
-    },
-
-    _initializeDeviceSelectionSettings(emitter) {
-        this.changeDevicesList([]);
-
-        $('#selectCamera').change(function () {
-            let cameraDeviceId = $(this).val();
-            if (cameraDeviceId !== Settings.getCameraDeviceId()) {
-                emitter.emit(UIEvents.VIDEO_DEVICE_CHANGED, cameraDeviceId);
-            }
-        });
-        $('#selectMic').change(function () {
-            let micDeviceId = $(this).val();
-            if (micDeviceId !== Settings.getMicDeviceId()) {
-                emitter.emit(UIEvents.AUDIO_DEVICE_CHANGED, micDeviceId);
-            }
-        });
-        $('#selectAudioOutput').change(function () {
-            let audioOutputDeviceId = $(this).val();
-            if (audioOutputDeviceId !== Settings.getAudioOutputDeviceId()) {
-                emitter.emit(
-                    UIEvents.AUDIO_OUTPUT_DEVICE_CHANGED, audioOutputDeviceId);
-            }
-        });
     },
 
     /**
@@ -286,91 +220,5 @@ export default {
      */
     isVisible () {
         return UIUtil.isVisible(document.getElementById("settings_container"));
-    },
-
-    /**
-     * Sets microphone's <select> element to select microphone ID from settings.
-     */
-    setSelectedMicFromSettings () {
-        $('#selectMic').val(Settings.getMicDeviceId());
-    },
-
-    /**
-     * Sets camera's <select> element to select camera ID from settings.
-     */
-    setSelectedCameraFromSettings () {
-        $('#selectCamera').val(Settings.getCameraDeviceId());
-    },
-
-    /**
-     * Sets audio outputs's <select> element to select audio output ID from
-     * settings.
-     */
-    setSelectedAudioOutputFromSettings () {
-        $('#selectAudioOutput').val(Settings.getAudioOutputDeviceId());
-    },
-
-    /**
-     * Change available cameras/microphones or hide selects completely if
-     * no devices available.
-     * @param {{ deviceId, label, kind }[]} devices list of available devices
-     */
-    changeDevicesList (devices) {
-        let $selectCamera= AJS.$('#selectCamera'),
-            $selectMic = AJS.$('#selectMic'),
-            $selectAudioOutput = AJS.$('#selectAudioOutput'),
-            $selectAudioOutputParent = $selectAudioOutput.parent();
-
-        let audio = devices.filter(device => device.kind === 'audioinput'),
-            video = devices.filter(device => device.kind === 'videoinput'),
-            audioOutput = devices
-                .filter(device => device.kind === 'audiooutput'),
-            selectedAudioDevice = audio.find(
-                d => d.deviceId === Settings.getMicDeviceId()) || audio[0],
-            selectedVideoDevice = video.find(
-                d => d.deviceId === Settings.getCameraDeviceId()) || video[0],
-            selectedAudioOutputDevice = audioOutput.find(
-                    d => d.deviceId === Settings.getAudioOutputDeviceId()),
-            videoPermissionGranted =
-                JitsiMeetJS.mediaDevices.isDevicePermissionGranted('video'),
-            audioPermissionGranted =
-                JitsiMeetJS.mediaDevices.isDevicePermissionGranted('audio');
-
-        $selectCamera
-            .html(generateDevicesOptions(
-                video,
-                selectedVideoDevice ? selectedVideoDevice.deviceId : '',
-                videoPermissionGranted))
-            .prop('disabled', !video.length || !videoPermissionGranted);
-
-        initSelect2($selectCamera);
-
-        $selectMic
-            .html(generateDevicesOptions(
-                audio,
-                selectedAudioDevice ? selectedAudioDevice.deviceId : '',
-                audioPermissionGranted))
-            .prop('disabled', !audio.length || !audioPermissionGranted);
-
-        initSelect2($selectMic);
-
-        if (JitsiMeetJS.mediaDevices.isDeviceChangeAvailable('output')) {
-            $selectAudioOutput
-                .html(generateDevicesOptions(
-                    audioOutput,
-                    selectedAudioOutputDevice
-                        ? selectedAudioOutputDevice.deviceId
-                        : 'default',
-                    videoPermissionGranted || audioPermissionGranted))
-                .prop('disabled', !audioOutput.length ||
-                    (!videoPermissionGranted && !audioPermissionGranted));
-            initSelect2($selectAudioOutput);
-
-            $selectAudioOutputParent.show();
-        } else {
-            $selectAudioOutputParent.hide();
-        }
-
-        APP.translation.translateElement($('#settings_container option'));
     }
 };
