@@ -1,27 +1,26 @@
-import { ReducerRegistry, setStateProperties } from '../redux';
+import { assign, ReducerRegistry } from '../redux';
 
-import {
-    HIDE_DIALOG,
-    OPEN_DIALOG
-} from './actionTypes';
+import { HIDE_DIALOG, OPEN_DIALOG } from './actionTypes';
 
 /**
- * Listen for actions which show or hide dialogs.
+ * Reduces redux actions which show or hide dialogs.
  *
- * @param {Object[]} state - Current state.
- * @param {Object} action - Action object.
- * @param {string} action.type - Type of action.
- * @returns {{}}
+ * @param {State} state - The current redux state.
+ * @param {Action} action - The redux action to reduce.
+ * @param {string} action.type - The type of the redux action to reduce..
+ * @returns {State} The next redux state that is the result of reducing the
+ * specified action.
  */
 ReducerRegistry.register('features/base/dialog', (state = {}, action) => {
     switch (action.type) {
     case HIDE_DIALOG:
-        return setStateProperties(state, {
+        return assign(state, {
             component: undefined,
             componentProps: undefined
         });
+
     case OPEN_DIALOG:
-        return setStateProperties(state, {
+        return assign(state, {
             component: action.component,
             componentProps: action.componentProps
         });
