@@ -1,8 +1,12 @@
+/* @flow */
+
+import { SET_CONFIG } from '../config';
+import { setLoggingConfig } from '../logging';
 import { PARTICIPANT_LEFT } from '../participants';
 import { MiddlewareRegistry } from '../redux';
 
 import { disposeLib, initLib, setWebRTCReady } from './actions';
-import { LIB_DID_INIT, LIB_INIT_ERROR, SET_CONFIG } from './actionTypes';
+import { LIB_DID_INIT, LIB_INIT_ERROR } from './actionTypes';
 import { WEBRTC_NOT_READY, WEBRTC_NOT_SUPPORTED } from './constants';
 
 /**
@@ -12,8 +16,8 @@ import { WEBRTC_NOT_READY, WEBRTC_NOT_SUPPORTED } from './constants';
  * lib-jitsi-meet, and initializes a new one with new config.
  *
  * @param {Store} store - Redux store.
- * @returns {Function}
  * @private
+ * @returns {Function}
  */
 MiddlewareRegistry.register(store => next => action => {
     switch (action.type) {
@@ -45,14 +49,14 @@ MiddlewareRegistry.register(store => next => action => {
  * specified action to the specified store.
  * @param {Action} action - The Redux action LIB_INIT_ERROR which is being
  * dispatched in the specified store.
+ * @private
  * @returns {Object} The new state that is the result of the reduction of the
  * specified action.
- * @private
  */
 function _libInitError(store, next, action) {
     const nextState = next(action);
 
-    const error = action.error;
+    const { error } = action;
 
     if (error) {
         let webRTCReady;
@@ -83,27 +87,34 @@ function _libInitError(store, next, action) {
  * specified action to the specified store.
  * @param {Action} action - The Redux action SET_CONFIG which is being
  * dispatched in the specified store.
+ * @private
  * @returns {Object} The new state that is the result of the reduction of the
  * specified action.
- * @private
  */
-function _setConfig(store, next, action) {
-    const { dispatch, getState } = store;
+function _setConfig({ dispatch, getState }, next, action) {
     const { initialized } = getState()['features/base/lib-jitsi-meet'];
 
     // XXX Since the config is changing, the library lib-jitsi-meet must be
-    // initialized again with the new config. Consequntly, it may need to be
+    // initialized again with the new config. Consequently, it may need to be
     // disposed of first.
     // TODO Currently, disposeLib actually does not dispose of lib-jitsi-meet
     // because lib-jitsi-meet does not implement such functionality.
     const disposeLibPromise
         = initialized ? dispatch(disposeLib()) : Promise.resolve();
 
-    // Let the new config into the Redux store (because initLib will read it
-    // from there).
-    const nextState = next(action);
+    disposeLibPromise.then(() => {
+        // Let the new config into the Redux store (because initLib will read it
+        // from there).
+        next(action);
 
-    disposeLibPromise.then(dispatch(initLib()));
+        // FIXME Obviously, the following is bad design. However, I'm currently
+        // introducing the features base/config and base/logging and I'm trying
+        // to minimize the scope of the changes while I'm attempting to preserve
+        // compatibility with the existing partially React-ified Web source code
+        // and what was already executing on React Native. Additionally, I do
+        // not care to load logging_config.js on React Native.
+        dispatch(setLoggingConfig(window.loggingConfig));
 
-    return nextState;
+        dispatch(initLib());
+    });
 }
