@@ -1,4 +1,5 @@
 export * from './actions';
 export * from './components';
+export * from './constants';
 
-import './reducer';
+import './middleware';
